New feature for name annotation editing
