+++ frontend/src/pages/annotation/components/asideAnnotation/AsideAnnotation.jsx
@@ -1,17 +1,4 @@
 import React, { useState, useEffect } from 'react';
-import file from '../../../projects/assets/fileIcon.png';
-import Modal from '../../../components/Modal';
-import { updateAnnotation } from '../utils/updateAnnotation/UpdateAnnotation';
-import {
-    Card,
-    CardContent,
-    CardDescription,
-    CardFooter,
-    CardHeader,
-    CardTitle,
-} from "@/components/ui/card"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
 import { Button } from "@/components/ui/button"
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Separator } from '@/components/ui/separator';
@@ -23,22 +10,18 @@ import {
     SheetTitle,
     SheetTrigger,
 } from "@/components/ui/sheet"
-import { ChevronRight } from 'lucide-react';
+import { ChevronRight, FileText } from 'lucide-react';
 
-const AsideAnnotation = ({ idProject, singleRequest, setSelectedAnnotation }) => {
-
-    // Recupera o valor de 'count' do localStorage ou define como 0 se não existir
-    const [modalIsOpen, setModalIsOpen] = useState(false);
-    const [newAnnotationName, setNewAnnotationName] = useState({
-        title: "",
-        id: "",
-    });
-
+const AsideAnnotation = ({
+    idProject,
+    singleRequest,
+    setSelectedAnnotation,
+    setNewAnnotationName,
+    handleUpdateName,
+    newAnnotationName,
+    setModalIsOpen
+}) => {
     const token = localStorage.getItem('token');
-
-    const handleUpdateName = async (id) => {
-        await updateAnnotation(token, newAnnotationName, id, idProject);
-    }
 
     return (
         <>
@@ -59,7 +42,7 @@ const AsideAnnotation = ({ idProject, singleRequest, setSelectedAnnotation }) =>
                                 singleRequest.annotations.map((annotation, index) => (
                                     <React.Fragment key={annotation.idAnnotation}>
                                         <div onClick={() => setSelectedAnnotation(annotation)} className="text-sm cursor-pointer list-none flex gap-2 items-center">
-                                            <img src={file} alt="file" />
+                                            <FileText width={20} />
                                             <li onDoubleClick={() => { setModalIsOpen(true); setNewAnnotationName({ title: annotation.title, id: annotation.idAnnotation }); }}>{annotation.title}</li>
                                         </div>
                                         <Separator className="my-2" />
@@ -75,30 +58,6 @@ const AsideAnnotation = ({ idProject, singleRequest, setSelectedAnnotation }) =>
                     </ScrollArea>
                 </SheetContent>
             </Sheet>
-
-            <Modal isOpen={modalIsOpen} onClose={(() => setModalIsOpen(false))}>
-                <Card>
-
-                    <CardHeader>
-                        <CardTitle>Update annotation name</CardTitle>
-                        <CardDescription>Change the name of your note.</CardDescription>
-                    </CardHeader>
-                    <CardContent>
-                        <div className="flex flex-col space-y-1.5">
-                            <Label htmlFor="name">Name</Label>
-                            <Input id="name" placeholder="Name of your project"
-                                value={newAnnotationName.title} // Ajuste para acessar a propriedade newTitle
-                                onChange={(e) => setNewAnnotationName({ ...newAnnotationName, title: e.target.value })} />
-                        </div>
-                    </CardContent>
-                    <CardFooter className="flex justify-between">
-                        <Button variant="outline" onClick={() => setModalIsOpen(false)}>Cancel</Button>
-                        <Button onClick={() => handleUpdateName(newAnnotationName.id)}>Save</Button>
-                    </CardFooter>
-
-                </Card>
-            </Modal>
-
         </>
     )
 }
